refactor `PodU16`, `PodU64`, `PodProofType`
