+++ zk-token-sdk/src/zk_token_proof_state.rs
@@ -1,5 +1,5 @@
 use {
-    crate::{zk_token_elgamal::pod::PodProofType, zk_token_proof_instruction::ProofType},
+    crate::{zk_token_elgamal::PodProofType, zk_token_proof_instruction::ProofType},
     bytemuck::{bytes_of, Pod, Zeroable},
     num_traits::ToPrimitive,
     solana_program::{
